+++ activesupport/CHANGELOG.md
@@ -1,11 +1,7 @@
-*   Allow assertionless tests to be reported.
+*   Warn on tests without assertions.
 
-    `ActiveSupport::TestCase` can be configured to report tests that do not run any assertions.
+    `ActiveSupport::TestCase` now warns when tests do not run any assertions.
     This is helpful in detecting broken tests that do not perform intended assertions.
-
-    ```ruby
-    config.active_support.assertionless_tests_behavior = :raise
-    ```
 
     *fatkodima*
